requirements.txt
discord[voice]
opuslib
wheel
ffmpeg
youtube_dl
PyNaCl
requests
datadog
google-api-python-client
wikipedia
googletrans==4.0.0-rc1
lyricsgenius
fuzzysearch
discord-pretty-help
lxml
langcodes
language_data
pytz
google-auth-httplib2
google-auth-oauthlib